+++ PongSDL/GameMath.h
@@ -2,11 +2,11 @@
 #define USE_GAMEMATH
 #include <math.h>
 struct Vector2 {
-    double x;
-    double y;
+    float x;
+    float y;
     Vector2() :x(0), y(0) {}
-    Vector2(double _x, double _y) : x(_x), y(_y) {}
-    Vector2(double _val) : x(_val), y(_val) {}
+    Vector2(float _x, float _y) : x(_x), y(_y) {}
+    Vector2(float _val) : x(_val), y(_val) {}
     Vector2& operator+=(const Vector2& _other) {
         this->x += _other.x;
         this->y += _other.y;
@@ -25,15 +25,15 @@ struct Vector2 {
         lhs -= rhs;
         return lhs;
     }
-    Vector2& operator*= (const double& other) {
+    Vector2& operator*= (const float& other) {
         this->x *= other;
         this->y *= other;
         return *this;
     }
-    friend Vector2 operator* (const double& lhs, Vector2 rhs) {
+    friend Vector2 operator* (const float& lhs, Vector2 rhs) {
         return rhs *= lhs;
     }
-    friend Vector2 operator* (Vector2 lhs, const double& rhs) {
+    friend Vector2 operator* (Vector2 lhs, const float& rhs) {
         return lhs *= rhs;
     }
     static Vector2 one() {
@@ -48,19 +48,19 @@ struct Vector2 {
     static Vector2 zero() {
         return Vector2();
     }
-    static double dot(Vector2 a, Vector2 b) {
+    static float dot(Vector2 a, Vector2 b) {
         return (a.x * b.x) + (a.y * b.y);
     }
-    double GetMagnitude() {
+    float GetMagnitude() {
         return sqrt(pow(x, 2)+ pow(y, 2));
     }
     void Normalise() {
-        double mag = GetMagnitude();
+        float mag = GetMagnitude();
         x = x / mag;
         y = y / mag;
     }
 };
-long double Lerp(long double a, long double b, long double fac);
+ float Lerp( float a,  float b,  float fac);
 #endif // !USE_GAMEMATH
